Plays with code, can't handle styling correctly, found libraries
It seems what I actually needed to search for was "fuzzy search",
not "lazy search". Once I did that I found libraries. I think
I should go with fuzzy.js and add a bit that prevents displaying
things that don't contain a letter in the search. Not sure how
advanced the score calculation is, but it seems good enough for
now.