Changes handling of tagName if empty string or if quotes
I figure they'll get a warning either way through the new
validation stuff. They don't need another, and the text was
inaccurate, because they won't just get a string of the result
with .toString(), they'll still have elements in there. I also
added a message specifically for that case because I can see
myself mistakenly sending that in in order to try to just get
the terms that matched.