Converting from using image objects to build grid to image nodes
Problem is that every time I set the grid, I currently make new
ImageChoice objects, which sets the image node's list node. When
I search the images for matches to queries, I use setGrid. That
means that all the hard work I did creating the list of matches
goes down the drain. The image choices need to be made at the
start and then their data should be used from then on.