Icon gets added to editor replacing text. Big bug?...
1) If new text is written right next to an existing icon and a new
icon is created for that text, the old icon gets replaced by the
new icon. That's not exactly a bug, since technically speaking
it's all just one word. Where <> represent text in the first icon:
<abit>oftext <- that's all actually one word, so it _should_
all be part of one icon. It's just weird when it happens.

2) Noticed that, once again, inserting images at the beginning
of the adder search bar, or before the first image if there are
images... or something... bugs come up with semicolons and with
invisible letters being matched in the search.