Fixes icons not showing in searchbar...
Reverts to text not being able to be in an icon. This had to
happen anyway - as it currently was, text could only be there if
it was last. There was no way to choose to use text in the middle.
A possible option is to have one choice box wich contains the text
being written, so it can be chosen as the contenst.
Issues:
- Width - what if the text width is wider than that of the image-
choice container?
- How to insert it and differentiate it from regular text as far
as the program goes?
Should probably write up an issue for this